// src/models/shareOrg.ts
export interface ShareOrg {
    id: number
    name: string
    address: any
    phoneNumber: any
    description: any
    pipedriveId: any
    linkedIn: any
    program_Name: any
    linked_In: any
    apartments: any
    condos: any
    senior_Housing: any
    student_Housing: any
    assisted_Living: any
    sfR_Portfolio: any
    mobile_Home_Park: any
    co_living: any
    office: any
    medical_Office: any
    manufacturing: any
    lt_Industrial: any
    cannabis: any
    retail: any
    hotel: any
    land: any
    self_storage: any
    religious: any
    hospital: any
    other: any
    leaseD_FEE: any
    tenancy: any
    single_Tenant_list: any
    single_Tenant__Min_Bond_Credit_Rating: any
    hotel__Flag_required: any
    hotel__Flag_list: any
    acquisition: any
    new_Development: any
    redevelopment: any
    refinance: any
    loan_Terms: any
    senior: any
    mezzanine: any
    preferred_Equity: any
    pace: any
    jV_Equity: any
    minimum_Loan_Size: any
    maximum_Loan_Size: any
    max_LTV: any
    max_LTC: any
    as_stabilized_LTV: any
    min_DSCR: any
    investor: any
    owner_Occupier: any
    target_States: any
    target_Counties: any
    sponsor_States: any
    international: any
    recourse: any
    corp_Guarantor_OK: any
    org_ID: any
    demo_Org_ID: any
    lender_Name: any
    loan_Program_Type: any
    contact_Email_Address: any
    startsTransparent: any
    lenderPays: any
    lenderFee: any
    disable_Auto_match: any
    contractStatus: any
    contact_Name: any
    contact_Number: any
    notes: any
    required_Files: any
    rate_Index: any
    minimum_Spread: any
    maximum_Spread: any
    typical_Amortization: any
    prepayment_Penalty: any
    last_Updated: any
    last_Updated_By: any
    capSnapId: any
    url: any
    states: any
    isActive: any
    lenderType: any
    summary: any
    organizationCategoryId: any
    status: any
    externalTeam: any
    internalTeam: any
    stakeholderId: any
    usage: any
    logoURL: any
    city: any
    state: any
    token:any
  }
  